+++ CHANGELOG.md
@@ -21,6 +21,9 @@
   * Make `Address` implement `Parcelable`
 * PayPalNativePayments
   * Bump `PayPal Native Checkout` to `1.0.0`
+* Breaking Changes
+  * CorePayments
+    * Remove `STAGING` constant from `Environment` enum
 
 ## 0.0.10 (2023-06-23)
 * Breaking Changes